@@ -4,27 +4,26 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=3.2.3 <4.0.0'
 
 dependencies:
   # presentation only deps
-  flutter_bloc: ^3.2.0
-  dartz: ^0.8.9
-  freezed_annotation: ^0.7.1
-  get_it: ^3.1.0
-  equatable: ^1.1.0
+  cupertino_icons: ^1.0.6
+  dartz: ^0.10.1
+  equatable: ^2.0.5
   flutter:
     sdk: flutter
+  flutter_bloc: ^8.1.3
+  freezed_annotation: ^2.4.1
+  get_it: ^7.6.7
 
-  cupertino_icons: ^0.1.2
 
 dev_dependencies:
-  mockito: ^4.1.1
-  lint: ^1.1.1
+  bloc_test: ^9.1.5
   build_runner:
-  freezed: ^0.9.2
-  flutter_test:
-    sdk: flutter
+  freezed: ^2.4.6
+  lint: ^2.3.0
+  mockito: ^5.4.4
 
 flutter:
   uses-material-design: true
