@@ -2,7 +2,7 @@ import 'package:equatable/equatable.dart';
 
 abstract class Failure extends Equatable {
   @override
-  List<Object> get props => null;
+  List<Object> get props => [];
 }
 
 class ServerFailure extends Failure {}
